src/pyproject.toml
### `pyproject.toml`
[project]
name = "aiops-agent"
version = "0.1.0"
description = "Prompt-driven AI Ops agent for Nephio and O-RAN"
authors = [{name = "Your Name"}]
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
aiops = "aiops_cli:app"

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.12.3"
pyyaml = "^6.0.2"
openai = "^1.40.0"

[build-system]
requires = ["setuptools", "wheel"]